Fixed BUG#25787 - java.util.Date should be serialized for  PreparedStatement.setObject().
We've added a new configuration option "treatUtilDateAsTimestamp", which is  false by default, 
as (1) We already had specific behavior to treat java.util.Date as a java.sql.Timestamp because 
it's useful to many folks, and (2) that behavior will very likely be required for drivers JDBC-post-4.0.

(this commit also fixes a missed merge of the above patch in 5.1).