Fixed BUG#51783 - Load-balanced connections could throw a SQLException incorrectly on commit() or rollback().  This was not caused by failures in commit or rollback, but rather by the possibility that the newly-selected physical connection was stale.  Added logic to catch and retry if this happens, up to the number of hosts specified for load-balancing.  Also added new property, loadBalanceValidateConnectionOnSwapServer, which controls whether to explicitly ping the selected host (otherwise, the host is presumed to be up, and will only be noticed if auto-commit or transaction isolation state needs to be set and fails).
Added loadBalancePingTimeout property to allow a specific timeout to be set
for each ping executed against the servers.  This ping is executed when the
physical connections are rebalanced (commit/rollback or communication exception),
or when a query starting with (exactly) "/* ping */" is executed.  The latter
causes each open underlying physical connection to be pinged.

Fixed BUG#51776 - Connection.rollback() could swallow exceptions incorrectly.