Bug #68916 closeOnCompletion doesn't work Bug #69746 ResultSet closed after Statement.close() when dontTrackOpenResources=true
StatementImpl handles closing itself when the last ResultSet is closed explicitely
and calls back to StatementImpl to remove itself from the set of open result sets.
Results from getGeneratedKeys() are considered part of this set of open result sets.

Implicit closing of result sets by StatementImpl was refactored to operate consistently
unless "dontTrackOpenResources" or "holdResultsOpenOverClose/holdResultsOpenOverStatementClose"
(stmt or conn level) is set.

Generated keys results were sometimes added to the set of open results. This is no longer done
and this result set is tracked separately. This resulted in a change to internal behaviour
checked in the test for Bug#44056. The test case was changed accordingly.

The description of "dontTrackOpenResources" was updated to reflect the new implications.

The "isClosed()" method was added to the "ResultSetInternalMethods" and moved from
JDBC4ResultSetImpl to ResultSetImpl.