Fix observed NPE in clearInputStream
Under load, we observe NPEs like this in our log files with the
8.0.12 driver:

Caused by: java.lang.NullPointerException: null
        at com.mysql.cj.protocol.a.NativeProtocol.clearInputStream(NativeProtocol.java:827)
        at com.mysql.cj.NativeSession.clearInputStream(NativeSession.java:335)
        at com.mysql.cj.jdbc.ServerPreparedStatement.serverPrepare(ServerPreparedStatement.java:659)
        at com.mysql.cj.jdbc.ServerPreparedStatement.<init>(ServerPreparedStatement.java:133)
        at com.mysql.cj.jdbc.ServerPreparedStatement.getInstance(ServerPreparedStatement.java:102)
        at com.mysql.cj.jdbc.ConnectionImpl.prepareStatement(ConnectionImpl.java:1649)
        at com.mysql.cj.jdbc.ConnectionImpl.prepareStatement(ConnectionImpl.java:1583)
[...]

This is dereferencing the socketConnection.getMysqlInput() input stream
as the this.socketConnection field only gets assigned once and never
overridden. This patch adds a null check to the while condition so that
a null value is not dereferenced.