Replace synchronized with ReentrantLock
Replace connectionMutex and synchronized blocks with a ReentrantLock. The synchronized blocks were removed and replaced by `lock.lock()` and `lock.unlock()`. This avoids that the carrier thread (OS thread) is pinned when running on virtual threads which were introduced as a preview feature in JDK 19.

The JEP that is being prepared for JDK 21:
https://openjdk.org/jeps/8303683

`There are two scenarios in which a virtual thread cannot be unmounted during blocking operations because it is pinned to its carrier:

 When it executes code inside a synchronized block or method, or
 When it executes a native method or a foreign function.`